Reworked Windowing System
Windowing is now handled in the DisplayManager without needed knowledge of the ModuleManager. Window class is now apart of ModuleInfo of a Module object so unecessary calls to run() of a Module were removed. The ModuleManager now has function to retrieve all WindowPtr's of all known modules. The displaying of a Window is only a test as of now.